Modified Files: 	sources/gkmod/klsupport_fwd.h
Changed PrimitiveRow (a list of block members) from size_t (= machine word)
to unsigned int (four bytes), saving four bytes per entry on a 64 bit machine.

+++ sources/gkmod/klsupport_fwd.h
@@ -27,7 +27,7 @@ namespace klsupport {
 class KLSupport;
 class KLSupport_new;
 
-typedef std::vector<size_t> PrimitiveRow;
+typedef std::vector<unsigned int> PrimitiveRow;
 
 }
